+++ src/DisassemblyPanel.tsx
@@ -15,12 +15,11 @@ const DisassemblyPanel: Component<IDisassemblyProp> = (prop: IDisassemblyProp) =
 
     hotkeys('ctrl+b', (e, _) => {
 		e.preventDefault()
-		toggleBreakpoint()
+		toggleBreakpoint(3) // 3 is middle element / program counter index
 	})
 
-    const toggleBreakpoint = () => {
-        // middle element is current instruction (TODO: refactor)
-        invoke('cmd_toggle_breakpoint', { address: instructions()[3][1] })
+    const toggleBreakpoint = (i: number) => {
+        invoke('cmd_toggle_breakpoint', { address: instructions()[i][1] })
     }
 
     onMount(async () => {
@@ -38,17 +37,17 @@ const DisassemblyPanel: Component<IDisassemblyProp> = (prop: IDisassemblyProp) =
                 <table onScroll={() => {}} class="font-mono w-full">
                     <thead>
                         <tr class="bg-gray-700">
-                            <td class="pl-2"></td>
+                            <td class="pl-2" colspan="2">BP</td>
                             <td class="pl-2">Address</td>
                             <td class="pl-6">Instruction</td>
                             <td class="pl-6">Assembly</td>
                         </tr>
                     </thead>
                     <tbody>
-                        {instructions().map((instruction: IDisassemblyInstruction, _) => {
+                        {instructions().map((instruction: IDisassemblyInstruction, i: number) => {
                             return (
                                 <tr class={instruction[1] === programCounter() ? 'bg-blue-900' : 'bg-gray-800'}>
-                                    <td class="pl-2 text-red-700">{ instruction[0] ? '◉' : ' ' }</td>
+                                    <td class="pl-2 text-red-700 cursor-pointer opacity-0 hover:opacity-50 active:opacity-100" classList={ {['opacity-100']: instruction[0]} } colspan="2" onClick={(_) => toggleBreakpoint(i)}>◉</td>
                                     <td class="pl-2">{instruction[1].toString(16).padStart(8, '0')}</td>
                                     <td class="pl-6">{instruction[2].toString(16).padStart(8, '0')}</td>
                                     <td class="pl-6">{instruction[3]}</td>